Move/Refactor ReflectionTest -> StackWalkerTest
sun.reflect.Reflection#getCallerClass is an internal API and is marked
for removal. This uses StackWalker to walk the stack and obtain the
caller class with StackWalker.StackFrame.getDeclaringClass instead.

The class is also being renamed from
'gov.nasa.jpf.test.vm.reflection.ReflectionTest' to
'gov.nasa.jpf.test.java.lang.StackWalkerTest' for consistency.

Fixes: #81
